requirements.txt
fastapi
uvicorn
pandas
python-multipart
thefuzz
python-Levenshtein
pytesseract
Pillow
opencv-python
spacy
